Move scout stuff into pkg/client/scout
