+++ pkg/client/scout/report.go
